Fix search for nova instance
When nova-multi-controller and neutron-server are deployed
on two different nodes, the neutron.conf was not properly updated
with the nova settings because it could't find the node. Search
for the instance (optional since normal deployment order currently
is neutron first, then nova) properly so that get_instance()
in the chef recipe can do its job.